requirements.txt
Pillow>=9.0
numpy>=1.22
pygltflib>=1.15
